Update patch.sh
Add debian support

+++ patch.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
-rsync -r ./lib/Virtualmin /usr/share/perl5/vendor_perl
+if [ -f /etc/lsb-release ]; then 
+  rsync -r ./lib/Virtualmin /usr/share/perl5
+elif [ -f /etc/redhat-release ]; then 
+  rsync -r ./lib/Virtualmin /usr/share/perl5/vendor_perl
+fi
+
+
